+++ vercel-ai-sdk/src/provider-response-provider.ts
@@ -6,9 +6,11 @@ import { Mem0ProviderSettings } from "./mem0-provider";
 import { CohereProviderSettings, createCohere } from "@ai-sdk/cohere";
 import { AnthropicProviderSettings, createAnthropic } from "@ai-sdk/anthropic";
 import { createGroq, GroqProviderSettings } from "@ai-sdk/groq";
+import { createDeepSeek, DeepSeekProviderSettings } from "@ai-sdk/deepseek";
+import { createOpenAICompatible, OpenAICompatibleProviderSettings} from "@ai-sdk/openai-compatible"
 
-export type Provider = ReturnType<typeof createOpenAI> | ReturnType<typeof createCohere> | ReturnType<typeof createAnthropic> | ReturnType<typeof createGroq> | any;
-export type ProviderSettings = OpenAIProviderSettings | CohereProviderSettings | AnthropicProviderSettings | GroqProviderSettings;
+export type Provider = ReturnType<typeof createOpenAI> | ReturnType<typeof createCohere> | ReturnType<typeof createAnthropic> | ReturnType<typeof createGroq> | ReturnType<typeof createDeepSeek> | any;
+export type ProviderSettings = OpenAIProviderSettings | CohereProviderSettings | AnthropicProviderSettings | GroqProviderSettings | DeepSeekProviderSettings | OpenAICompatibleProviderSettings;
 
 class Mem0AITextGenerator {
     provider: Provider;
@@ -53,8 +55,22 @@ class Mem0AITextGenerator {
                     ...provider_config,
                 });
                 break;
+            case "deepseek":
+                this.provider = createDeepSeek({
+                    apiKey: config?.apiKey,
+                    ...provider_config,
+                });
+                break;
+            case "aliyun":
+                this.provider = createOpenAICompatible({
+                    apiKey: config?.apiKey,
+                    baseURL: "https://dashscope.aliyuncs.com/compatible-mode/v1",
+                    name: 'aliyun',
+                    ...provider_config,
+                });
+                break;
             default:
-                throw new Error("Invalid provider");
+                throw new Error("Unsupported provider");
         }
         this.model = model;
         this.provider_config = provider_config;